依存関係調整
